fix(triton): type-103 MoE crashed on fp16/bf16 activations
The decoded weight tile stayed float32 because codes was cast to float32
before multiplying by the scale, so the product promoted back regardless
of the scale cast. tl.dot rejects mismatched operand dtypes.

Real models run bf16, so this failed on every non-fp32 model. It was
invisible because the parity suite only ever built float32 activations.

Extend that suite to fp16/bf16 so the gap cannot reopen. float16 needs
the activations rescaled first: random packed codes with live scales give
weight magnitudes far above a trained model, and the gate*up product
makes the pipeline quadratic in x, pushing intermediates past float16 max.

Same bug class as the type-100 and type-101 dtype fixes.

+++ vllm_gguf_plugin/triton/fused_moe/standard_quant/q8_0_rocmfpx.py
@@ -113,7 +113,10 @@ def _q8_0_rocmfpx_moe_kernel(
         w_tile = codes * scale[:, :, None]
         w_tile = tl.reshape(w_tile, (BLOCK_N, BLOCK_K_BLOCKS * BLOCK_SIZE))
 
-        acc = tl.dot(x_tile, tl.trans(w_tile), acc=acc)
+        # codes is float32, so codes * scale promotes back to float32 even
+        # though scale was cast to x_dtype. tl.dot requires matching operand
+        # dtypes, so cast explicitly rather than relying on the product.
+        acc = tl.dot(x_tile, tl.trans(w_tile.to(x_dtype)), acc=acc)
 
     y_ptrs = y_ptr + offs_output[:, None] * stride_ym + offs_n[None, :] * stride_yn
     y_mask = token_mask[:, None] & (offs_n[None, :] < n)
